requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
Flask==1.1.2
Flask-SocketIO==5.0.1
Flask-Sockets==0.2.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
prompt-toolkit==3.0.18
pyflakes==2.3.1
Pygments==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pyvmomi==7.0.2
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
